+++ SondasEmMarte/src/comandos/Verificacao.java
@@ -1,0 +1,5 @@
+package comandos;
+
+public class Verificacao {
+
+}
