+++ SondasEmMarte/src/testesondas/Teste.java
@@ -1,0 +1,19 @@
+package testesondas;
+
+import plato.Plato;
+import sondas.Sonda;
+
+public class Teste{
+	private static Plato temp1 = new Plato();
+	public static void main(String[] args) {
+		
+
+	}
+	
+	public static Plato getTemp1() {
+		return temp1;
+	}
+	public static void setTemp1(Plato temp1) {
+		Teste.temp1 = temp1;
+	}
+}
